+++ VulkanMemoryAllocator/changelog.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## WIP
+- Bump VMA, Adds virtual allocation functionality
 
 ## [0.7.2] - 2021-10-14
 - Bump VMA, documentation changes and internal improvements